Reorganize project structure into core and ui modules
