+++ src/Rpg/gamengine/RPGEngine.h
@@ -14,7 +14,7 @@
 #include "../npcs/GarrickStone.h"
 #include "../npcs/VincentHale.h"
 #include "../npcs/SeraphinaLumeris.h"
-#include "../../TimeSystem.h"
+#include "../../core/TimeSystem.h"
 #include "../menues/BankMenu.h"
 #include "../inventory/items/Wood.h"
 #include "../inventory/items/TowerBlueprint.h"
@@ -33,7 +33,7 @@
 #include "../map/buildings/Barrier.h"
 #include "../map/buildings/Bed.h"
 #include "../map/buildings/Chest.h"
-#include "../../TransitionSystem.h"
+#include "../../core/TransitionSystem.h"
 #include "../map/entities/EntityFactory.h"
 #include "../map/zones/ZoneManager.h"
 #include "../map/entities/WaypointManager.h"
