+++ src/Rpg/skillTree/SkillTreeMenu.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <SFML/Graphics.hpp>
 #include "SkillTree.h"
-#include "../../TowerDefense/menues/Button.h"
+#include "../../ui/Button.h"
 
 class SkillTreeMenu {
 public:
