+++ src/Source.cpp
@@ -1,4 +1,4 @@
-#include "GameManager.h"
+#include "core/GameManager.h"
 
 int main() {
     GameManager gameManager;
