+++ src/TowerDefense/menues/GameOverMenu.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <SFML/Graphics.hpp>
-#include "Button.h"
+#include "../../ui/Button.h"
 #include <vector>
 
 class GameEngine;
