+++ src/core/Animation.cpp
