+++ src/core/Animation.h
