+++ src/core/GameManager.cpp
