+++ src/core/ResourceManager.cpp
