+++ src/core/ResourceManager.h
