+++ src/core/TransitionSystem.cpp
