+++ src/core/TransitionSystem.h
